requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
python-multipart
python-dotenv
pydantic-settings
pydantic[email]